refactor: move base project to separated folder, to have better visibility
