+++ Sorter-Solver/SorterSolver.h
