+++ Sorter-Solver/SortingProblem.cpp
