+++ Sorter-Solver/SortingProblemState.h
