+++ SorterSolver.sln
@@ -3,14 +3,11 @@ Microsoft Visual Studio Solution File, Format Version 12.00
 # Visual Studio Version 16
 VisualStudioVersion = 16.0.32602.291
 MinimumVisualStudioVersion = 10.0.40219.1
-Project("{8BC9CEB8-8B4A-11D0-8D11-00A0C91BC942}") = "SorterSolver", "SorterSolver.vcxproj", "{36A91A6C-1A60-4F47-A12F-B0B914B6224B}"
-EndProject
 Project("{8BC9CEB8-8B4A-11D0-8D11-00A0C91BC942}") = "Sorter-Unit-Test", "Sorter-Unit-Test\Sorter-Unit-Test.vcxproj", "{C0EE9448-96F7-4F6F-9F38-357B530AE9CD}"
 EndProject
 Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Solution Items", "Solution Items", "{B3F54850-4D2A-46CF-B39E-58A99EBEA132}"
-	ProjectSection(SolutionItems) = preProject
-		CMakeLists.txt = CMakeLists.txt
-	EndProjectSection
+EndProject
+Project("{8BC9CEB8-8B4A-11D0-8D11-00A0C91BC942}") = "SorterSolver", "Sorter-Solver\SorterSolver.vcxproj", "{36A91A6C-1A60-4F47-A12F-B0B914B6224B}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -20,14 +17,6 @@ Global
 		Release|x86 = Release|x86
 	EndGlobalSection
 	GlobalSection(ProjectConfigurationPlatforms) = postSolution
-		{36A91A6C-1A60-4F47-A12F-B0B914B6224B}.Debug|x64.ActiveCfg = Debug|x64
-		{36A91A6C-1A60-4F47-A12F-B0B914B6224B}.Debug|x64.Build.0 = Debug|x64
-		{36A91A6C-1A60-4F47-A12F-B0B914B6224B}.Debug|x86.ActiveCfg = Debug|Win32
-		{36A91A6C-1A60-4F47-A12F-B0B914B6224B}.Debug|x86.Build.0 = Debug|Win32
-		{36A91A6C-1A60-4F47-A12F-B0B914B6224B}.Release|x64.ActiveCfg = Release|x64
-		{36A91A6C-1A60-4F47-A12F-B0B914B6224B}.Release|x64.Build.0 = Release|x64
-		{36A91A6C-1A60-4F47-A12F-B0B914B6224B}.Release|x86.ActiveCfg = Release|Win32
-		{36A91A6C-1A60-4F47-A12F-B0B914B6224B}.Release|x86.Build.0 = Release|Win32
 		{C0EE9448-96F7-4F6F-9F38-357B530AE9CD}.Debug|x64.ActiveCfg = Debug|x64
 		{C0EE9448-96F7-4F6F-9F38-357B530AE9CD}.Debug|x64.Build.0 = Debug|x64
 		{C0EE9448-96F7-4F6F-9F38-357B530AE9CD}.Debug|x86.ActiveCfg = Debug|Win32
@@ -36,6 +25,14 @@ Global
 		{C0EE9448-96F7-4F6F-9F38-357B530AE9CD}.Release|x64.Build.0 = Release|x64
 		{C0EE9448-96F7-4F6F-9F38-357B530AE9CD}.Release|x86.ActiveCfg = Release|Win32
 		{C0EE9448-96F7-4F6F-9F38-357B530AE9CD}.Release|x86.Build.0 = Release|Win32
+		{36A91A6C-1A60-4F47-A12F-B0B914B6224B}.Debug|x64.ActiveCfg = Debug|x64
+		{36A91A6C-1A60-4F47-A12F-B0B914B6224B}.Debug|x64.Build.0 = Debug|x64
+		{36A91A6C-1A60-4F47-A12F-B0B914B6224B}.Debug|x86.ActiveCfg = Debug|Win32
+		{36A91A6C-1A60-4F47-A12F-B0B914B6224B}.Debug|x86.Build.0 = Debug|Win32
+		{36A91A6C-1A60-4F47-A12F-B0B914B6224B}.Release|x64.ActiveCfg = Release|x64
+		{36A91A6C-1A60-4F47-A12F-B0B914B6224B}.Release|x64.Build.0 = Release|x64
+		{36A91A6C-1A60-4F47-A12F-B0B914B6224B}.Release|x86.ActiveCfg = Release|Win32
+		{36A91A6C-1A60-4F47-A12F-B0B914B6224B}.Release|x86.Build.0 = Release|Win32
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
